# openapiSpec.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Migracion CRM by Johny Marin Gutierrez",
    "version": "0.1.0"
  },
  "paths": {
    "/api/cities": {
      "get": {
        "summary": "Get Cities",
        "operationId": "get_cities_api_cities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Cities Api Cities Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityPydantic"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create City",
        "operationId": "create_city_api_cities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CityInPydantic"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityPydantic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cities/{city_id}": {
      "get": {
        "summary": "Read City",
        "operationId": "read_city_api_cities__city_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "City Id",
              "type": "integer"
            },
            "name": "city_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityPydantic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update City",
        "operationId": "update_city_api_cities__city_id__put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "City Id",
              "type": "integer"
            },
            "name": "city_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CityOptPydantic"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityPydantic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete City",
        "operationId": "delete_city_api_cities__city_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "City Id",
              "type": "integer"
            },
            "name": "city_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/": {
      "get": {
        "summary": "Root",
        "operationId": "root_api__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/hello/{name}": {
      "get": {
        "summary": "Say Hello",
        "operationId": "say_hello_api_hello__name__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            },
            "name": "name",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CityInPydantic": {
        "title": "CityInPydantic",
        "required": [
          "timezone"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "default": "CityIn"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        }
      },
      "CityOptPydantic": {
        "title": "CityOptPydantic",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "default": "CityOpt"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        }
      },
      "CityPydantic": {
        "title": "CityPydantic",
        "required": [
          "id",
          "timezone"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string",
            "default": "City"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}